GuideList

+++ src/main/java/com/proyecto/flowmanagement/backend/service/IGuideService.java
@@ -9,6 +9,6 @@ public interface IGuideService {
 	Guide add(Guide guide);
 	Guide update(Guide guide);
 	List<Guide> getAll();
-	Guide getById(Integer id);
-	void delete(Integer id);
+	Guide getById(Long id);
+	void delete(Long id);
 }
